requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aioredis==1.3.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
bardapi==0.1.21
beautifulsoup4==4.12.2
breadability==0.1.20
Brotli==1.0.9
cachetools==4.2.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
deep-translator==1.11.1
Deprecated==1.2.13
diskcache==5.6.1
dnspython==2.3.0
docopt==0.6.2
duckduckgo-search==3.8.5
email-validator==1.3.1
fastapi==0.92.0
frozenlist==1.3.3
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.48.2
h11==0.14.0
h2==4.1.0
hiredis==2.2.2
hpack==4.0.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langdetect==1.0.9
libretranslatepy==2.1.1
limits==2.8.0
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.32
nltk==3.8.1
numpy==1.24.3
openai==0.27.4
orjson==3.8.6
packaging==22.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pydantic==1.10.5
python-dotenv==1.0.0
python-multipart==0.0.5
PyYAML==6.0
regex==2023.3.23
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
slowapi==0.1.7
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.4
starlette==0.25.0
sumy==0.11.0
tqdm==4.65.0
translate==3.6.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.14.1
yarl==1.8.2